# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SatQuery AI",
    "description": "Agentic vision-language analysis for satellite imagery.\n\n**Schema version 1.0 \u2014 frozen.** Fields may be added within `1.0`; renaming,\nremoving or retyping any field requires a `2.0` bump. See `DOCS/API_CONTRACT.md`.\n\nPhase 0: every endpoint returns schema-valid fixtures.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Queue an analysis and watch it stream",
        "description": "Accept the same body as ``/v1/analyze`` and return immediately.\n\nThe returned ``job_id`` *is* the eventual ``trace_id``, so the client can\naddress the trace before the pipeline has produced it.",
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_v1_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Poll one job's progress",
        "description": "Return the job's current stage, progress and \u2014 once finished \u2014 its result.\n\nA job the store no longer holds \u2014 evicted, or from before a restart \u2014 is\nanswered from its persisted trace when one exists: the trace store is the\ndurable half by design, and a client reconnecting after a server restart\nshould get its answer rather than a 404 it cannot tell from \"never existed\".",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel a running job",
        "description": "Ask a running job to stop. Best-effort, and honest about what that means.\n\nThe tool step in flight finishes \u2014 a worker thread cannot be killed \u2014 and no\nfurther step starts. The job then ends with an ``error`` event carrying\n``JOB_CANCELLED``, so an open events stream closes the way it would for any\nother failure. ``202`` because the stop is requested, not yet done.",
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or evicted job."
          },
          "409": {
            "description": "The job has already finished.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Server-sent progress events for one job",
        "description": "Stream the job's events, replaying everything that already happened.\n\nA subscriber that connects late \u2014 the normal case, since the client needs the\n``job_id`` from the 202 first \u2014 receives the buffered history before the live\ntail, so the DAG can always be drawn from the ``plan`` event. One that\n*reconnects* sends the last ``id:`` it saw and receives only the rest.",
        "operationId": "job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resume after this event id; the query form of Last-Event-ID.",
              "title": "After"
            },
            "description": "Resume after this event id; the query form of Last-Event-ID."
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream. Event order is guaranteed: queued -> stage -> plan -> interleaved step_started/step_completed/artifact -> answer_delta* -> done|error. `done` and `error` are terminal and mutually exclusive.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or evicted job."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "maxLength": 32,
            "minLength": 32,
            "pattern": "^[0-9a-f]{32}$",
            "title": "Trace Id"
          },
          "answer": {
            "$ref": "#/components/schemas/Answer"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRef"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "compatibility": {
            "$ref": "#/components/schemas/CompatibilityReport"
          },
          "resolved_task": {
            "$ref": "#/components/schemas/ResolvedTask"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuditTrace"
              },
              {
                "type": "null"
              }
            ],
            "description": "None when options.include_trace is false."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "answer",
          "confidence",
          "compatibility",
          "resolved_task"
        ],
        "title": "AnalyzeResponse",
        "description": "``POST /v1/analyze`` 200 body.\n\n``compatibility`` and ``resolved_task`` are duplicated at the top level so the\nUI can render them without walking into the trace."
      },
      "Answer": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "uncited_numeric_spans": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncited Numeric Spans",
            "description": "Numbers the CitationValidator could not resolve. Non-empty is a visible honesty signal \u2014 the frontend must render it."
          },
          "generator": {
            "type": "string",
            "title": "Generator",
            "description": "e.g. 'vlm_change_vqa@1.0.0+adapter:sq-lora-v3'."
          },
          "template_fallback": {
            "type": "boolean",
            "title": "Template Fallback",
            "description": "True when a deterministic template wrote the text.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "generator"
        ],
        "title": "Answer",
        "description": "The synthesised natural-language answer plus its honesty signals."
      },
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "http_status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Http Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint",
            "description": "An actionable next step."
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "http_status",
          "message"
        ],
        "title": "ApiError",
        "description": "The error payload (API_CONTRACT \u00a76).\n\n``message`` and ``hint`` are user-facing and safe to display verbatim: never\nleak file paths or stack traces into either."
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "description": "Wire envelope for :class:`ApiError` \u2014 every non-2xx body has this shape."
      },
...
      "Body_create_job_v1_jobs_post": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images",
            "description": "1-2 files. GeoTIFF preferred; PNG/JPEG accepted."
          },
          "query": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "description": "1-1000 chars, UTF-8."
          },
          "options": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options",
            "description": "JSON, see AnalyzeOptions."
          }
        },
        "type": "object",
        "required": [
          "images",
          "query"
        ],
        "title": "Body_create_job_v1_jobs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "maxLength": 32,
            "minLength": 32,
            "pattern": "^[0-9a-f]{32}$",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          },
          "poll_url": {
            "type": "string",
            "title": "Poll Url",
            "description": "e.g. '/v1/jobs/b3f1...'."
          },
          "events_url": {
            "type": "string",
            "title": "Events Url",
            "description": "SSE endpoint, e.g. '/v1/jobs/b3f1.../events'."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "poll_url",
          "events_url"
        ],
        "title": "JobAccepted",
        "description": "``POST /v1/jobs`` 202 body (API_CONTRACT \u00a74.2).\n\n``job_id`` *is* the eventual ``trace_id``: the id is minted before the\npipeline runs, so a client can address the trace before it exists."
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "maxLength": 32,
            "minLength": 32,
            "pattern": "^[0-9a-f]{32}$",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "description": "queued|ingesting|validating|rendering|planning|executing|aggregating|done."
          },
          "step": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Step"
          },
          "total_steps": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Steps"
          },
          "pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Pct"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyzeResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "status",
          "stage",
          "step",
          "total_steps",
          "pct",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "description": "``GET /v1/jobs/{job_id}`` 200 body (API_CONTRACT \u00a74.3).\n\n``result`` is populated only once ``status`` is ``succeeded``, ``error``\nonly once it is ``failed``; they are mutually exclusive, mirroring the\n``done``/``error`` terminal SSE events."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}